backend/requirements.txt
# Core framework
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
pydantic>=2.9.0
pydantic-settings>=2.5.0
python-dotenv>=1.0.0
python-multipart>=0.0.9

# LLM / LangGraph
openai>=1.50.0
langgraph>=0.2.0
langchain-core>=0.3.0

# Database
sqlalchemy>=2.0.0
psycopg2-binary>=2.9.0

# Vector store (RAG)
chromadb>=0.5.0

# Charts / data
matplotlib>=3.9.0
pandas>=2.2.0
numpy>=1.26.0

# Document parsing
pypdf>=4.0.0
python-docx>=1.1.0
